+++ sdk/search/azure-search-documents/azure/search/documents/_service/_models.py
@@ -3,10 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-from ._generated.models import (
-    Analyzer,
-    Tokenizer
-)
+from ._generated.models import Analyzer, Tokenizer
 
 
 class PatternAnalyzer(Analyzer):
