+++ sdk/search/azure-search-documents/tests/async_tests/recordings/test_service_live_async.test_create_datasource_async.yaml
@@ -11,37 +11,30 @@ interactions:
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - D59D8B92A259BB430B2DD58824AC5800
+      - 705C140806CA0B3C9E38BE08AF7425C2
     method: POST
     uri: https://search386b1565.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://search386b1565.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E76B55FC94A6\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
+      string: '{"@odata.context":"https://search386b1565.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E8A051BE1734\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
     headers:
       cache-control: no-cache
       content-length: '370'
       content-type: application/json; odata.metadata=minimal
-      date: Thu, 23 Apr 2020 09:47:28 GMT
-      elapsed-time: '36'
-      etag: W/"0x8D7E76B55FC94A6"
+      date: Fri, 24 Apr 2020 22:39:16 GMT
+      elapsed-time: '43'
+      etag: W/"0x8D7E8A051BE1734"
       expires: '-1'
       location: https://search386b1565.search.windows.net/datasources('sample-datasource')?api-version=2019-05-06-Preview
       odata-version: '4.0'
       pragma: no-cache
       preference-applied: odata.include-annotations="*"
-      request-id: 71366868-8547-11ea-a573-2816a845e8c6
+      request-id: 6bd3982a-867c-11ea-8426-8c8590507855
       strict-transport-security: max-age=15724800; includeSubDomains
     status:
       code: 201
       message: Created
-    url: !!python/object/new:yarl.URL
-      state: !!python/tuple
-      - !!python/object/new:urllib.parse.SplitResult
-        - https
-        - search386b1565.search.windows.net
-        - /datasources
-        - api-version=2019-05-06-Preview
-        - ''
+    url: https://search386b1565.search.windows.net/datasources?api-version=2019-05-06-Preview
 version: 1
