+++ sdk/search/azure-search-documents/tests/async_tests/recordings/test_service_live_async.test_delete_datasource_async.yaml
@@ -11,87 +11,73 @@ interactions:
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - B9AD3FA1A5A00F28630A84A8C57C2595
+      - 7A89725BBC59BE289D165493C43F6C66
     method: POST
     uri: https://search38471564.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://search38471564.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E76B7020FF97\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
+      string: '{"@odata.context":"https://search38471564.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E8A0605A970A\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
     headers:
       cache-control: no-cache
       content-length: '370'
       content-type: application/json; odata.metadata=minimal
-      date: Thu, 23 Apr 2020 09:48:12 GMT
-      elapsed-time: '52'
-      etag: W/"0x8D7E76B7020FF97"
+      date: Fri, 24 Apr 2020 22:39:39 GMT
+      elapsed-time: '56'
+      etag: W/"0x8D7E8A0605A970A"
       expires: '-1'
       location: https://search38471564.search.windows.net/datasources('sample-datasource')?api-version=2019-05-06-Preview
       odata-version: '4.0'
       pragma: no-cache
       preference-applied: odata.include-annotations="*"
-      request-id: 8b696eba-8547-11ea-a514-2816a845e8c6
+      request-id: 7a71a03e-867c-11ea-8426-8c8590507855
       strict-transport-security: max-age=15724800; includeSubDomains
     status:
       code: 201
       message: Created
-    url: !!python/object/new:yarl.URL
-      state: !!python/tuple
-      - !!python/object/new:urllib.parse.SplitResult
-        - https
-        - search38471564.search.windows.net
-        - /datasources
-        - api-version=2019-05-06-Preview
-        - ''
+    url: https://search38471564.search.windows.net/datasources?api-version=2019-05-06-Preview
 - request:
     body: null
     headers:
       Accept:
       - application/json;odata.metadata=minimal
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - B9AD3FA1A5A00F28630A84A8C57C2595
+      - 7A89725BBC59BE289D165493C43F6C66
     method: GET
     uri: https://search38471564.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://search38471564.search.windows.net/$metadata#datasources","value":[{"@odata.etag":"\"0x8D7E76B7020FF97\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}]}'
+      string: '{"@odata.context":"https://search38471564.search.windows.net/$metadata#datasources","value":[{"@odata.etag":"\"0x8D7E8A0605A970A\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}]}'
     headers:
       cache-control: no-cache
       content-encoding: gzip
-      content-length: '366'
+      content-length: '365'
       content-type: application/json; odata.metadata=minimal
-      date: Thu, 23 Apr 2020 09:48:12 GMT
-      elapsed-time: '22'
+      date: Fri, 24 Apr 2020 22:39:39 GMT
+      elapsed-time: '24'
       expires: '-1'
       odata-version: '4.0'
       pragma: no-cache
       preference-applied: odata.include-annotations="*"
-      request-id: 8b8d7e18-8547-11ea-8d15-2816a845e8c6
+      request-id: 7a88e582-867c-11ea-8426-8c8590507855
       strict-transport-security: max-age=15724800; includeSubDomains
       vary: Accept-Encoding
     status:
       code: 200
       message: OK
-    url: !!python/object/new:yarl.URL
-      state: !!python/tuple
-      - !!python/object/new:urllib.parse.SplitResult
-        - https
-        - search38471564.search.windows.net
-        - /datasources
-        - api-version=2019-05-06-Preview
-        - ''
+    url: https://search38471564.search.windows.net/datasources?api-version=2019-05-06-Preview
 - request:
     body: null
     headers:
       Accept:
       - application/json;odata.metadata=minimal
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - B9AD3FA1A5A00F28630A84A8C57C2595
+      - 7A89725BBC59BE289D165493C43F6C66
     method: DELETE
     uri: https://search38471564.search.windows.net/datasources('sample-datasource')?api-version=2019-05-06-Preview
   response:
@@ -99,32 +85,25 @@ interactions:
       string: ''
     headers:
       cache-control: no-cache
-      date: Thu, 23 Apr 2020 09:48:12 GMT
-      elapsed-time: '26'
+      date: Fri, 24 Apr 2020 22:39:39 GMT
+      elapsed-time: '41'
       expires: '-1'
       pragma: no-cache
-      request-id: 8b97b236-8547-11ea-87a8-2816a845e8c6
+      request-id: 7a91254e-867c-11ea-8426-8c8590507855
       strict-transport-security: max-age=15724800; includeSubDomains
     status:
       code: 204
       message: No Content
-    url: !!python/object/new:yarl.URL
-      state: !!python/tuple
-      - !!python/object/new:urllib.parse.SplitResult
-        - https
-        - search38471564.search.windows.net
-        - /datasources('sample-datasource')
-        - api-version=2019-05-06-Preview
-        - ''
+    url: https://search38471564.search.windows.net/datasources('sample-datasource')?api-version=2019-05-06-Preview
 - request:
     body: null
     headers:
       Accept:
       - application/json;odata.metadata=minimal
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - B9AD3FA1A5A00F28630A84A8C57C2595
+      - 7A89725BBC59BE289D165493C43F6C66
     method: GET
     uri: https://search38471564.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
@@ -135,24 +114,17 @@ interactions:
       content-encoding: gzip
       content-length: '202'
       content-type: application/json; odata.metadata=minimal
-      date: Thu, 23 Apr 2020 09:48:12 GMT
+      date: Fri, 24 Apr 2020 22:39:39 GMT
       elapsed-time: '5'
       expires: '-1'
       odata-version: '4.0'
       pragma: no-cache
       preference-applied: odata.include-annotations="*"
-      request-id: 8ba2572e-8547-11ea-b58f-2816a845e8c6
+      request-id: 7a9b8098-867c-11ea-8426-8c8590507855
       strict-transport-security: max-age=15724800; includeSubDomains
       vary: Accept-Encoding
     status:
       code: 200
       message: OK
-    url: !!python/object/new:yarl.URL
-      state: !!python/tuple
-      - !!python/object/new:urllib.parse.SplitResult
-        - https
-        - search38471564.search.windows.net
-        - /datasources
-        - api-version=2019-05-06-Preview
-        - ''
+    url: https://search38471564.search.windows.net/datasources?api-version=2019-05-06-Preview
 version: 1
