+++ sdk/search/azure-search-documents/tests/async_tests/recordings/test_service_live_async.test_get_datasource_async.yaml
@@ -11,77 +11,63 @@ interactions:
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - 51E1A53A641E5EE7069C83D5F23E7C4C
+      - 9F98C05C3DB836017EA5D1617BDE33AF
     method: POST
     uri: https://searchfa0d1431.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://searchfa0d1431.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E76B7D1B6295\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
+      string: '{"@odata.context":"https://searchfa0d1431.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E8A067C498B3\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
     headers:
       cache-control: no-cache
       content-length: '370'
       content-type: application/json; odata.metadata=minimal
-      date: Thu, 23 Apr 2020 09:48:33 GMT
-      elapsed-time: '38'
-      etag: W/"0x8D7E76B7D1B6295"
+      date: Fri, 24 Apr 2020 22:39:52 GMT
+      elapsed-time: '51'
+      etag: W/"0x8D7E8A067C498B3"
       expires: '-1'
       location: https://searchfa0d1431.search.windows.net/datasources('sample-datasource')?api-version=2019-05-06-Preview
       odata-version: '4.0'
       pragma: no-cache
       preference-applied: odata.include-annotations="*"
-      request-id: 98587b54-8547-11ea-8bd6-2816a845e8c6
+      request-id: 81dc3c58-867c-11ea-8426-8c8590507855
       strict-transport-security: max-age=15724800; includeSubDomains
     status:
       code: 201
       message: Created
-    url: !!python/object/new:yarl.URL
-      state: !!python/tuple
-      - !!python/object/new:urllib.parse.SplitResult
-        - https
-        - searchfa0d1431.search.windows.net
-        - /datasources
-        - api-version=2019-05-06-Preview
-        - ''
+    url: https://searchfa0d1431.search.windows.net/datasources?api-version=2019-05-06-Preview
 - request:
     body: null
     headers:
       Accept:
       - application/json;odata.metadata=minimal
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - 51E1A53A641E5EE7069C83D5F23E7C4C
+      - 9F98C05C3DB836017EA5D1617BDE33AF
     method: GET
     uri: https://searchfa0d1431.search.windows.net/datasources('sample-datasource')?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://searchfa0d1431.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E76B7D1B6295\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
+      string: '{"@odata.context":"https://searchfa0d1431.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E8A067C498B3\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
     headers:
       cache-control: no-cache
       content-encoding: gzip
-      content-length: '359'
+      content-length: '358'
       content-type: application/json; odata.metadata=minimal
-      date: Thu, 23 Apr 2020 09:48:33 GMT
+      date: Fri, 24 Apr 2020 22:39:52 GMT
       elapsed-time: '8'
-      etag: W/"0x8D7E76B7D1B6295"
+      etag: W/"0x8D7E8A067C498B3"
       expires: '-1'
       odata-version: '4.0'
       pragma: no-cache
       preference-applied: odata.include-annotations="*"
-      request-id: 9888909e-8547-11ea-9e77-2816a845e8c6
+      request-id: 81f2b212-867c-11ea-8426-8c8590507855
       strict-transport-security: max-age=15724800; includeSubDomains
       vary: Accept-Encoding
     status:
       code: 200
       message: OK
-    url: !!python/object/new:yarl.URL
-      state: !!python/tuple
-      - !!python/object/new:urllib.parse.SplitResult
-        - https
-        - searchfa0d1431.search.windows.net
-        - /datasources('sample-datasource')
-        - api-version=2019-05-06-Preview
-        - ''
+    url: https://searchfa0d1431.search.windows.net/datasources('sample-datasource')?api-version=2019-05-06-Preview
 version: 1
