+++ sdk/search/azure-search-documents/tests/recordings/test_service_live.test_create_datasource.yaml
@@ -15,14 +15,14 @@ interactions:
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - 6EE1AFB6E75621F4C0D8091DB40240CE
+      - 5944A0DC7FD275C43424B38DC65B1807
     method: POST
     uri: https://search51c7106b.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://search51c7106b.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E76B0EE7CAE6\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
+      string: '{"@odata.context":"https://search51c7106b.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E8A02E96F7B0\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
     headers:
       cache-control:
       - no-cache
@@ -31,11 +31,11 @@ interactions:
       content-type:
       - application/json; odata.metadata=minimal
       date:
-      - Thu, 23 Apr 2020 09:45:29 GMT
+      - Fri, 24 Apr 2020 22:38:16 GMT
       elapsed-time:
-      - '59'
+      - '29'
       etag:
-      - W/"0x8D7E76B0EE7CAE6"
+      - W/"0x8D7E8A02E96F7B0"
       expires:
       - '-1'
       location:
@@ -47,7 +47,7 @@ interactions:
       preference-applied:
       - odata.include-annotations="*"
       request-id:
-      - 29fadd68-8547-11ea-946b-2816a845e8c6
+      - 48ac5f12-867c-11ea-8426-8c8590507855
       strict-transport-security:
       - max-age=15724800; includeSubDomains
     status:
