+++ sdk/search/azure-search-documents/tests/recordings/test_service_live.test_delete_datasource.yaml
@@ -15,14 +15,14 @@ interactions:
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - 8D1DA77C1D33A3A2C2E202F91D866E5C
+      - 0854D7F39E2712F07D51D39AD7B4BCB7
     method: POST
     uri: https://search51a9106a.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://search51a9106a.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E76B266466FE\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
+      string: '{"@odata.context":"https://search51a9106a.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E8A03C7859A1\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
     headers:
       cache-control:
       - no-cache
@@ -31,11 +31,11 @@ interactions:
       content-type:
       - application/json; odata.metadata=minimal
       date:
-      - Thu, 23 Apr 2020 09:46:08 GMT
+      - Fri, 24 Apr 2020 22:38:39 GMT
       elapsed-time:
-      - '51'
+      - '64'
       etag:
-      - W/"0x8D7E76B266466FE"
+      - W/"0x8D7E8A03C7859A1"
       expires:
       - '-1'
       location:
@@ -47,7 +47,7 @@ interactions:
       preference-applied:
       - odata.include-annotations="*"
       request-id:
-      - 419d323e-8547-11ea-a215-2816a845e8c6
+      - 5687b41a-867c-11ea-8426-8c8590507855
       strict-transport-security:
       - max-age=15724800; includeSubDomains
     status:
@@ -63,14 +63,14 @@ interactions:
       Connection:
       - keep-alive
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - 8D1DA77C1D33A3A2C2E202F91D866E5C
+      - 0854D7F39E2712F07D51D39AD7B4BCB7
     method: GET
     uri: https://search51a9106a.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://search51a9106a.search.windows.net/$metadata#datasources","value":[{"@odata.etag":"\"0x8D7E76B266466FE\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}]}'
+      string: '{"@odata.context":"https://search51a9106a.search.windows.net/$metadata#datasources","value":[{"@odata.etag":"\"0x8D7E8A03C7859A1\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}]}'
     headers:
       cache-control:
       - no-cache
@@ -79,9 +79,9 @@ interactions:
       content-type:
       - application/json; odata.metadata=minimal
       date:
-      - Thu, 23 Apr 2020 09:46:08 GMT
+      - Fri, 24 Apr 2020 22:38:39 GMT
       elapsed-time:
-      - '18'
+      - '39'
       expires:
       - '-1'
       odata-version:
@@ -91,7 +91,7 @@ interactions:
       preference-applied:
       - odata.include-annotations="*"
       request-id:
-      - 41d079fa-8547-11ea-9f01-2816a845e8c6
+      - 56a96858-867c-11ea-8426-8c8590507855
       strict-transport-security:
       - max-age=15724800; includeSubDomains
       vary:
@@ -111,9 +111,9 @@ interactions:
       Content-Length:
       - '0'
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - 8D1DA77C1D33A3A2C2E202F91D866E5C
+      - 0854D7F39E2712F07D51D39AD7B4BCB7
     method: DELETE
     uri: https://search51a9106a.search.windows.net/datasources('sample-datasource')?api-version=2019-05-06-Preview
   response:
@@ -123,15 +123,15 @@ interactions:
       cache-control:
       - no-cache
       date:
-      - Thu, 23 Apr 2020 09:46:08 GMT
+      - Fri, 24 Apr 2020 22:38:39 GMT
       elapsed-time:
-      - '54'
+      - '44'
       expires:
       - '-1'
       pragma:
       - no-cache
       request-id:
-      - 41dff18c-8547-11ea-8d51-2816a845e8c6
+      - 56b5e736-867c-11ea-8426-8c8590507855
       strict-transport-security:
       - max-age=15724800; includeSubDomains
     status:
@@ -147,9 +147,9 @@ interactions:
       Connection:
       - keep-alive
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - 8D1DA77C1D33A3A2C2E202F91D866E5C
+      - 0854D7F39E2712F07D51D39AD7B4BCB7
     method: GET
     uri: https://search51a9106a.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
@@ -163,9 +163,9 @@ interactions:
       content-type:
       - application/json; odata.metadata=minimal
       date:
-      - Thu, 23 Apr 2020 09:46:08 GMT
+      - Fri, 24 Apr 2020 22:38:39 GMT
       elapsed-time:
-      - '6'
+      - '8'
       expires:
       - '-1'
       odata-version:
@@ -175,7 +175,7 @@ interactions:
       preference-applied:
       - odata.include-annotations="*"
       request-id:
-      - 41f5654a-8547-11ea-a7a9-2816a845e8c6
+      - 56c2b128-867c-11ea-8426-8c8590507855
       strict-transport-security:
       - max-age=15724800; includeSubDomains
       vary:
