+++ sdk/search/azure-search-documents/tests/recordings/test_service_live.test_get_datasource.yaml
@@ -15,14 +15,14 @@ interactions:
       Content-Type:
       - application/json
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - 8CD7256132FD7881C2091E30F89DA85F
+      - 8F7E74DA6770FF09E7893CF94EE4ECAC
     method: POST
     uri: https://search22270f37.search.windows.net/datasources?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://search22270f37.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E76B3349695F\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
+      string: '{"@odata.context":"https://search22270f37.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E8A0435D5C22\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
     headers:
       cache-control:
       - no-cache
@@ -31,11 +31,11 @@ interactions:
       content-type:
       - application/json; odata.metadata=minimal
       date:
-      - Thu, 23 Apr 2020 09:46:30 GMT
+      - Fri, 24 Apr 2020 22:38:51 GMT
       elapsed-time:
-      - '49'
+      - '51'
       etag:
-      - W/"0x8D7E76B3349695F"
+      - W/"0x8D7E8A0435D5C22"
       expires:
       - '-1'
       location:
@@ -47,7 +47,7 @@ interactions:
       preference-applied:
       - odata.include-annotations="*"
       request-id:
-      - 4e608d6e-8547-11ea-8181-2816a845e8c6
+      - 5d6c2fea-867c-11ea-8426-8c8590507855
       strict-transport-security:
       - max-age=15724800; includeSubDomains
     status:
@@ -63,14 +63,14 @@ interactions:
       Connection:
       - keep-alive
       User-Agent:
-      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Windows-10-10.0.17763-SP0)
+      - azsdk-python-search-documents/1.0.0b3 Python/3.7.3 (Darwin-19.3.0-x86_64-i386-64bit)
       api-key:
-      - 8CD7256132FD7881C2091E30F89DA85F
+      - 8F7E74DA6770FF09E7893CF94EE4ECAC
     method: GET
     uri: https://search22270f37.search.windows.net/datasources('sample-datasource')?api-version=2019-05-06-Preview
   response:
     body:
-      string: '{"@odata.context":"https://search22270f37.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E76B3349695F\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
+      string: '{"@odata.context":"https://search22270f37.search.windows.net/$metadata#datasources/$entity","@odata.etag":"\"0x8D7E8A0435D5C22\"","name":"sample-datasource","description":null,"type":"azureblob","subtype":null,"credentials":{"connectionString":null},"container":{"name":"searchcontainer","query":null},"dataChangeDetectionPolicy":null,"dataDeletionDetectionPolicy":null}'
     headers:
       cache-control:
       - no-cache
@@ -79,11 +79,11 @@ interactions:
       content-type:
       - application/json; odata.metadata=minimal
       date:
-      - Thu, 23 Apr 2020 09:46:30 GMT
+      - Fri, 24 Apr 2020 22:38:51 GMT
       elapsed-time:
-      - '11'
+      - '10'
       etag:
-      - W/"0x8D7E76B3349695F"
+      - W/"0x8D7E8A0435D5C22"
       expires:
       - '-1'
       odata-version:
@@ -93,7 +93,7 @@ interactions:
       preference-applied:
       - odata.include-annotations="*"
       request-id:
-      - 4eb64b08-8547-11ea-a135-2816a845e8c6
+      - 5d8e1a88-867c-11ea-8426-8c8590507855
       strict-transport-security:
       - max-age=15724800; includeSubDomains
       vary:
